Compilation working again.

+++ asteroids/types.cpp
@@ -1,12 +1,8 @@
 #include "types.h"
 
 namespace asteroids {
-
-  
-  float Constants::ASTEROID_SCALE = 0.3f;
-  float Constants::SHIP_SCALE = 0.3f;
-  int Constants::BULLET_WIDTH = 2;
-    int Constants::THETA_INCR = 3;
-};
-
-
+float Constants::ASTEROID_SCALE = 0.3f;
+float Constants::SHIP_SCALE = 0.3f;
+int Constants::BULLET_WIDTH = 2;
+float Constants::THETA_INCR = 3;
+}; // namespace asteroids
